src/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==11.0.3
google-cloud-speech==2.21.0
google-cloud-storage==2.13.0
pydantic==2.4.2
python-dotenv==1.0.0
numpy==1.24.3
sounddevice==0.4.6
soundfile==0.12.1

# 追加するパッケージ
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
